rename util since specific for ip allow list entries
